+++ changelog.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Breaking
 - upgrade: update spring-boot from v2.7.5 to v3.0.0
 
+### Added
+- ability to provide request path and timeout for health check
+
 ## [0.6.0] - 2022-11-23
 
 ### Breaking